Main Merger
This is going to be the current uploaded website to main, the main changes needed to be made is adding more pop-ups to the subsystems page. we will need to fix the footer on about & subsystems page, fix responsiveness on subsystems page and sponsors page as well as update sponsor package.